Fix in-memory write buffer overflow causing hang on large UPDATE workloads
The write buffer growth function (csGrowWriteBuf) used 32-bit int for size
tracking and a doubling strategy. For in-memory databases with large write
workloads (e.g., 10K updates on indexed columns), the buffer exceeded 1GB
and the next doubling overflowed int to a negative value, causing an
infinite loop in the while(nNew < nRequired) growth loop.

Fix: widen nWriteBuf, nWriteBufAlloc, and nCommittedWriteBuf from int to
i64, use sqlite3_realloc64 for the allocation, and switch from pure
doubling to 50% growth once past 64MB to reduce over-allocation.

+++ src/chunk_store.h
@@ -98,12 +98,12 @@ struct ChunkStore {
   int nPending;
   int nPendingAlloc;
   u8 *pWriteBuf;             /* Buffer for pending chunk data */
-  int nWriteBuf;
-  int nWriteBufAlloc;
+  i64 nWriteBuf;
+  i64 nWriteBufAlloc;
 
   u8 readOnly;               /* True if opened read-only */
   u8 isMemory;               /* True if this is an in-memory store */
-  int nCommittedWriteBuf;    /* Bytes in pWriteBuf belonging to committed chunks */
+  i64 nCommittedWriteBuf;    /* Bytes in pWriteBuf belonging to committed chunks */
 };
 
 /* Open or create a chunk store at the given path */
